refactor(Flowr): mode deep link related classes to internal
