+++ flowr/src/main/java/com/fueled/flowr/internal/FlowrConfig.java
@@ -1,4 +1,4 @@
-package com.fueled.flowr;
+package com.fueled.flowr.internal;
 
 /**
  * Created by julienFueled on 5/23/17.
